+++ Audio.h
@@ -1,0 +1,12 @@
+#pragma once
+namespace FreeTouchDeck
+{
+
+  enum class Sounds
+  {
+    GOING_TO_SLEEP,
+    BEEP,
+    STARTUP
+  };
+  void HandleAudio(Sounds sound);
+}
