+++ Console.h
@@ -1,0 +1,4 @@
+#pragma once
+namespace FreeTouchDeck {
+    void processSerial();
+}
