+++ MenuNavigation.h
@@ -3,17 +3,19 @@
 #include "Menu.h"
 #include "globals.hpp"
 namespace FreeTouchDeck {
-    extern void LoadAllMenus();
-    extern void LoadSystemMenus();
-    extern bool SetActiveScreen(const char * name);
-    extern bool ScreenLock(TickType_t xTicksToWait) ;
-    extern void ScreenUnlock() ;
-    extern char *MenusToJson(bool withSystem = false);
-    extern Menu *GetActiveScreen();
-    extern Menu *GetScreen(const char *name, bool lock=true);
-    extern bool SaveFullFormat();
-    extern Menu *GetLatchScreen(FTAction *action);
-    extern bool LoadFullFormat(const char * fileName);
-    extern bool LoadFullFormat();
+    void LoadAllMenus();
+    void LoadSystemMenus();
+    bool SetActiveScreen(const char * name);
+    bool ScreenLock(TickType_t xTicksToWait) ;
+    void ScreenUnlock() ;
+    char *MenusToJson(bool withSystem = false);
+    Menu *GetActiveScreen();
+    Menu *GetScreen(const char *name, bool lock=true);
+    bool SaveFullFormat();
+    Menu *GetLatchScreen(FTAction *action);
+    bool LoadFullFormat(const char * fileName);
+    bool LoadFullFormat();
+    void handleDisplay(bool pressed, uint16_t t_x, uint16_t t_y);
+    
 }
